override MPLPlot._get_colours()
